+++ R/ggplot_hbar_fun.R
@@ -30,7 +30,6 @@ theme_hbar <-
           face = "plain",
           hjust = 0.5
         ),
-        plot.title.position = "plot",
         plot.caption = element_text(
           family = font_family,
           colour = "#323232",
@@ -38,7 +37,6 @@ theme_hbar <-
           face = "plain",
           hjust = 0.99
         ),
-        plot.caption.position = "plot",
         plot.margin = margin(
           t = 5,
           l = 5,